refactor: 将重复的验证码逻辑封装进hook # Reviewed, transaction id: 21896
